Fix signed/unsigned compare warnings

+++ RTClib.cpp
@@ -194,7 +194,7 @@ DateTime::DateTime(uint32_t t) {
   uint8_t leap;
   for (yOff = 0;; ++yOff) {
     leap = yOff % 4 == 0;
-    if (days < 365 + leap)
+    if (days < 365U + leap)
       break;
     days -= 365 + leap;
   }
@@ -448,7 +448,7 @@ char *DateTime::toString(char *buffer) {
     }
   }
 
-  for (int i = 0; i < strlen(buffer) - 1; i++) {
+  for (size_t i = 0; i < strlen(buffer) - 1; i++) {
     if (buffer[i] == 'h' && buffer[i + 1] == 'h') {
       if (!apTag) { // 24 Hour Mode
         buffer[i] = '0' + hh / 10;
